cycles through job now.
main changes:
Latches for the heater and mix moved to lad2. Added a state 0 for the repeating cycles that starts the cycle over without the safe start and unlatches flags.  Added interlocks for heater because it comes on when there is nothing in batch otherwise. Removed reset drain line because it locked up the drain. 
Notes:
Currently the code will not stop at the job number; instead it makes batches until you pause and reset. The code is fairly messy at the moment and will need to be cleaned up.